pyproject.toml
[tool.poetry]
name = "logo-detector"
version = "0.1.0"
description = ""
authors = ["Saman Moeinsadat <moein.saman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
torchvision = "^0.3.0"
torch = "^1.1"
Pillow = "^6.1"
requests = "^2.22"
flask = "^1.1"
flask_cors = "^3.0"
clipper_admin = "^0.4.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pylint = "^2.3"
ipdb = "^0.12.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
